requirements.txt
coverage==7.11.0
iniconfig==2.3.0
packaging==25.0
pluggy==1.6.0
pygments==2.19.2
pytest==8.4.2
pytest-cov==7.0.0
python-dateutil==2.9.0
vendas-cli @ file://.